Moved local notification to types
