+++ src/app/services/firebase/cloud-messaging/local-notification.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { ILocalNotification, LocalNotifications } from '@ionic-native/local-notifications/ngx';
-import { LocalNotification } from './local-notification';
+import { LocalNotification } from './types/local-notification';
 
 @Injectable({
   providedIn: 'root',
